TWCCManager: split stats and twcc
